feat: add-mock-workers-cli-http-stability-smoke-002 - Prove /status and /work remain readable across the CLI lifecycle
